+++ codex/librarian/bulk_import/aggregate_metadata.py
@@ -49,8 +49,11 @@ def _get_path_metadata(path):
         # faster than getting the cover later in another thread.
         # Writing the image in another thread is faster than writing
         # it here.
-        task = ImageComicCoverCreateTask(True, path, cover_path, car.cover_image_data)
-        LIBRARIAN_QUEUE.put_nowait(task)
+        if car.cover_image_data:
+            task = ImageComicCoverCreateTask(
+                True, path, cover_path, car.cover_image_data
+            )
+            LIBRARIAN_QUEUE.put_nowait(task)
         _clean_md(md)
         group_tree = []
         for group_cls in BROWSER_GROUPS:
@@ -171,7 +174,7 @@ def get_aggregate_metadata(library, all_paths):
     all_failed_imports = {}
     total_paths = len(all_paths)
 
-    LOG.debug(f"Aggregating metadata in {library.path}...")
+    LOG.verbose(f"Aggregating metadata in {library.path}...")  # type: ignore
     last_log_time = time.time()
     for num, path in enumerate(all_paths):
         path = str(path)
@@ -191,12 +194,12 @@ def get_aggregate_metadata(library, all_paths):
 
         now = time.time()
         if now - last_log_time > LOG_EVERY:
-            LOG.debug(f"Aggregated {num}/{total_paths} comics")
+            LOG.verbose(f"Aggregated {num}/{total_paths} comics")  # type: ignore
             last_log_time = now
 
     all_fks["comic_paths"] = set(all_mds.keys())
 
     _bulk_update_or_create_failed_imports(library.pk, all_failed_imports)
 
-    LOG.debug(f"Aggregated metadata from {len(all_mds)} comics.")
+    LOG.verbose(f"Aggregated metadata from {len(all_mds)} comics.")  # type: ignore
     return all_mds, all_m2m_mds, all_fks