+++ run.sh
@@ -3,6 +3,6 @@
 THIS_DIR="$(dirname "$0")"
 cd "$THIS_DIR" || exit 1
 export OBJC_DISABLE_INITIALIZE_FORK_SAFETY=YES
-export PYTHONDONTWRITEBYTECODE=$DEBUG
+export PYTHONDONTWRITEBYTECODE="$DEBUG"
 export PYTHONPATH="$PYTHONPATH:$THIS_DIR"
 poetry run python3 ./codex/run.py